rename rule file so it's clear it's not to be used in case someone tries to play funny buggers
